requirements.txt
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
Django==2.0.6
django-cors-headers==2.2.0
django-extensions==2.0.7
django-utils==0.0.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docopt==0.6.2
ecdsa==0.13
ForgeryPy==0.1
future==0.16.0
idna==2.7
itypes==1.1.0
jedi==0.12.0
Jinja2==2.10
MarkupSafe==1.0
parso==0.2.1
Pillow==5.1.0
prompt-toolkit==1.0.15
ptpython==0.41
pyasn1==0.4.4
pycparser==2.18
Pygments==2.2.0
PyJWT==1.6.4
python-decouple==3.1
python-jose==3.0.0
pytz==2018.4
requests==2.19.1
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
wcwidth==0.1.7
